update: Zonkylla saves locally the transactions
Zonkylla now create './zonkylla.db' file where it saves all
transactions obtained from the server.

This is only core functionality. It could be nice to ask server only
for pieces of data from the date the last update was done. Another
point is that we want to save datetime data as INT (unix timestamp) or
REAL (julian date), not as the TEXT.

The question is if we would like to save enums data as TEXT or we
would like to implement the logic in a code... or we would like to
create separate table for it.

Resolves #20