Refactors _request method to AbstractClient
Moves method _request from OAuthClient to AbstractClient.

This includes defining of _client abstractmethod on AbstractClient and
appropriate methods on OAuthClient and Client which returns object
request call is made with.

Also implements _additional_params method which returns dict of params passed to
request method on _client object. Now it is used with OAuthClient to pass
client_id and client_secret.