database: Table of updates
We can see two internal value in z_internals table.

The first is db_version which corresponds with version of database. If code
is newer, user will be prompt to delete entire database, so clear update
can be done.

The second is updated. Everytime we update data in database from
zonky.cz this value is added. It prevents against huge network traffic.

Resolves #76